video 61

+++ src/services/OrderService.js
@@ -68,7 +68,7 @@ const createOrder = (newOrder) => {
         }
     })
 }
-const getAllOrderDetails = (id) => {
+const getAllOrderDetails = (id, data) => {
     return new Promise(async (resolve, reject) => {
         try {
             const order = await Order.find({ user: id })
@@ -112,20 +112,52 @@ const getDetailsOrder = (id) => {
         }
     })
 }
-const cancelOrder = (id) => {
+const cancelOrder = (id, data) => {
     return new Promise(async (resolve, reject) => {
         try {
-            const order = await Order.findByIdAndDelete(id)
-            // kiem tra user co trung khong
-            if (order === null) {
+            let order = []
+            const promises = data.map(async (order) => {
+                const productData = await Product.findOneAndUpdate(
+                    {
+                        _id: order.product,
+                        selled: { $gte: order.amount }
+                    },
+                    {
+                        $inc: {
+                            countInStock: +order.amount,
+                            selled: -order.amount
+                        }
+                    },
+                    { new: true }
+                )
+                console.log('productData', productData)
+                if (productData) {
+                    order = await Order.findByIdAndDelete(id)
+                    if (order === null) {
+                        resolve({
+                            status: 'ERR',
+                            message: 'The order is not defined'
+                        })
+                    }
+                } else {
+                    return {
+                        status: 'OK',
+                        message: 'ERR',
+                        id: order.product
+                    }
+                }
+            })
+            const results = await Promise.all(promises)
+            const newData = results && results.filter((item) => item)
+            if (newData.length) {
                 resolve({
-                    status: "OK",
-                    message: "the order is not defined"
+                    status: 'ERR',
+                    message: `San pham voi id${newData.join(',')} khong ton tai`
                 })
             }
             resolve({
-                status: "OK",
-                message: "SUCCESS",
+                status: 'OK',
+                message: 'success',
                 data: order
             })
         }
